refactoring : 마이페이지 세부기능 추상화
